build+docs: mypy strict gate, expanded ruff rules, ADRs, architecture, interview prep
Subagent #2 (mypy/ruff hardening):
- Expanded ruff lint set from {I, E, F, W, B, UP} to also include
  {ARG, SIM, RET, RUF, S} (unused args, simplifiable code, return
  quality, ruff-specific, security/bandit subset).
- 5 per-file ignores with justification comments (FastAPI Depends,
  Result Union for 3.10 TypeVar, _SessionLocal assert narrowing,
  ehr_client._c assert narrowing, en-dash in voice copy, pytest assert,
  eval setup-lambda signatures).
- Added mypy --strict hook to .pre-commit and a Type check step to
  .github/workflows/ci.yml.
- 17 source files now pass mypy --strict cleanly. Fixed type holes in
  llm.py (generate signature), ehr_client.py (Optional narrowing),
  dispatcher.py, tools.py, timing.py, runner.py, repository.py.

Subagent #3 (docs + ADRs + reviewer-readiness):
- docs/adr/{001..003}.md — 3 ADRs (hybrid FSM, separate EHR process,
  paired state+judge eval) in Context/Decision/Consequences/
  Alternatives format.
- docs/architecture.md — process topology + FSM state graph diagrams.
- docs/interview-notes.md — 10 likely interview questions + crisp
  answers for the candidate to review pre-interview.
- README.md — Windows uv-direct command column per make target, env-var
  table, eval CLI exit-code table, pointers to new docs.

Tests: 46/46 pass. mypy --strict: 17 files clean.

+++ src/prosper/observability/timing.py
@@ -18,7 +18,7 @@ def _percentile(values: list[float], pct: float) -> float:
     if not values:
         return 0.0
     s = sorted(values)
-    k = int(round((pct / 100) * (len(s) - 1)))
+    k = round((pct / 100) * (len(s) - 1))
     return s[max(0, min(len(s) - 1, k))]
